♻️ refactor : 폴더명과 겹치는 파일이름 index.tsx로 변경
