+++ src/app/(after-login)/dashboard/[id]/edit/loading.tsx
@@ -1,5 +1,5 @@
 import { Page, PageInner } from '@/components/layout/Page';
-import { CardSkeleton } from '@/components/ui/Card/Card';
+import { CardSkeleton } from '@/components/ui/Card';
 import { GoBackSkeleton } from '@/components/ui/Link/GoBackLink';
 
 export default function loading() {
