+++ src/app/error.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import React from 'react';
 import LogoCi from '@/assets/images/logo_ci.svg';
 import LogoBi from '@/assets/images/logo_bi.svg';
-import Button from '@/components/ui/Button/Button';
+import Button from '@/components/ui/Button';
 import Link from 'next/link';
 import axios from 'axios';
