+++ src/components/dashboard/DetailModify.tsx
@@ -3,12 +3,12 @@
 import { Controller, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useUpdateDashboard } from '@/apis/dashboards/queries';
-import { Card, CardTitle } from '@/components/ui/Card/Card';
+import { Card, CardTitle } from '@/components/ui/Card';
 import { Input } from '@/components/ui/Field';
 import ColorPicker from '@/components/ui/Chip/ColorPicker';
 import { Dashboard, dashboardFormSchema, DashboardFormType } from '@/apis/dashboards/types';
 import { getErrorMessage } from '@/utils/errorMessage';
-import Button from '@/components/ui/Button/Button';
+import Button from '@/components/ui/Button';
 import useAlert from '@/hooks/useAlert';
 import { useRouter } from 'next/navigation';
