+++ src/components/dashboard/EditCommentModal.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { forwardRef, useEffect, useState } from 'react';
-import { Modal, ModalContent, ModalFooter, ModalHandle, ModalHeader } from '../ui/Modal/Modal';
+import { Modal, ModalContent, ModalFooter, ModalHandle, ModalHeader } from '../ui/Modal';
 import { Textarea } from '../ui/Field';
-import Button from '../ui/Button/Button';
+import Button from '../ui/Button';
 
 interface EditCommentModalProps {
   initialContent: string;
