+++ src/components/dashboard/SidebarDashboardList.tsx
@@ -6,7 +6,7 @@ import Image from 'next/image';
 import { cn } from '@/utils/helper';
 import { useDashboardsQuery } from '@/apis/dashboards/queries';
 import PaginationControls from '@/components/pagination/PaginationControls';
-import Dot from '@/components/ui/Dot/Dot';
+import Dot from '@/components/ui/Dot';
 import crown from '@/assets/icons/crown.svg';
 import menu from '@/assets/icons/hamburger.svg';
 import { useParams } from 'next/navigation';
