+++ src/components/profile/PasswordEdit.tsx
@@ -5,8 +5,8 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Input } from '@/components/ui/Field/Input';
 import { passwordSchema, PutPasswordFormData } from '@/apis/auth/types';
 import useAlert from '@/hooks/useAlert';
-import { Card, CardTitle } from '@/components//ui/Card/Card';
-import Button from '@/components/ui/Button/Button';
+import { Card, CardTitle } from '@/components/ui/Card';
+import Button from '@/components/ui/Button';
 import { getErrorMessage } from '@/utils/errorMessage';
 import { usePutPassword } from '@/apis/auth/queries';
