+++ src/components/profile/ProfileEdit.tsx
@@ -8,8 +8,8 @@ import { UpdateUserForm, updateUserFormSchema, User } from '@/apis/users/types';
 import { useCreateProfileImage, useUpdateUser } from '@/apis/users/queries';
 import { Input } from '@/components/ui/Field/Input';
 import { ImageUpload } from '@/components/ui/Field/ImageUpload';
-import Button from '@/components/ui/Button/Button';
-import { Card, CardTitle } from '@/components//ui/Card/Card';
+import Button from '@/components/ui/Button';
+import { Card, CardTitle } from '@/components/ui/Card';
 import { getErrorMessage } from '@/utils/errorMessage';
 
 type ProfileEditProps = {
